+++ silverShelter/src/main/resources/templates/medicalCenter/AICaregiverMatchingService.html
@@ -1,7 +1,7 @@
 
 <h1 class="surveyFormTit" id="surveyFormTit1">Golden Prestige 의료센터에는 전문 요양사들이 등록되어있습니다</h1>
     <form id="surveyForm" action="/medicalCenter/careGiversResult" method="POST">
-        <h1 id="surveyFormTit" class="surveyFormTit">AI 요양사 매칭 사전 설문 조사</AI></h1>
+        <h1 id="surveyFormTit" class="surveyFormTit">요양사 매칭 사전 설문 조사</AI></h1>
         <p>선호하는 요양사 성별:</p>
         <input type="radio" id="male" name="gender" value="male">
         <label for="male">남자</label><br>
