+++ silverShelter/src/main/resources/templates/medicalCenter/careGiverMatching.html
@@ -29,15 +29,15 @@
     <th:block id="caregiverInfo" th:if="${session.loginMember.caregiversName != null}">
         <div id="caregiverInfo1">
             <h3 id="caregiverInfoTit">현재 요양사 정보 : <span th:text="${session.loginMember.caregiversName}" /> 요양사</h3>
-            <h4 id="caregiverInfoTit1">요양사 변경을 원한다면 AI 요양사 매칭 사전 설문조사를 다시 작성해주세요</h4>
-            <button id="AIBtn">AI 요양사 매칭 사전 설문조사 바로가기</button>
+            <h4 id="caregiverInfoTit1">요양사 변경을 원한다면 요양사 매칭 사전 설문조사를 다시 작성해주세요</h4>
+            <button id="AIBtn">요양사 매칭 사전 설문조사 바로가기</button>
         </div>
         <div id="AIMatching" style="display: none;" >
-        <th:block th:replace="~{medicalCenter/AICaregiverMatchingService}"></th:block>
+        <th:block th:replace="~{medicalCenter/CaregiverMatchingService}"></th:block>
         </div>
     </th:block>
 
     <th:block th:if="${session.loginMember.caregiversName == null}">
-        <th:block th:replace="~{medicalCenter/AICaregiverMatchingService}"></th:block>
+        <th:block th:replace="~{medicalCenter/CaregiverMatchingService}"></th:block>
         </th:block>
 </th:block>
